making all components accessible from components dir
