+++ app/components/index.ts
@@ -1,3 +1,11 @@
 export * from './ui/index.ts'
 export * from './layout/index.ts'
 export * from './templates/index.ts'
+
+export * from './confetti.tsx'
+export * from './error-boundary.tsx'
+export * from './floating-toolbar.tsx'
+export * from './progress-bar.tsx'
+export * from './search-bar.tsx'
+export * from './spacer.tsx'
+export * from './toaster.tsx'
