+++ app/components/layout/sidebar-list.tsx
@@ -1,7 +1,7 @@
 import { NavLink } from '@remix-run/react'
+import { Icon } from '#app/components/index.ts'
 import { cn } from '#app/utils/misc.tsx'
 import { type OwnerListData } from '#app/utils/user.ts'
-import { Icon } from '../ui/icon.tsx'
 import { ListItem } from './list-item.tsx'
 import { List } from './list.tsx'
