+++ app/components/ui/custom/index.ts
@@ -1,3 +1,5 @@
+export * from './checkbox-field.tsx'
 export * from './errors.tsx'
 export * from './field.tsx'
+export * from './status-button.tsx'
 export * from './textarea-field.tsx'
