+++ app/components/ui/index.ts
@@ -1,1 +1,2 @@
+export * from './custom/index.ts'
 export * from './primitives/index.ts'
