+++ app/components/ui/primitives/index.ts
@@ -1,2 +1,8 @@
 export * from './button.tsx'
 export * from './checkbox.tsx'
+export * from './dropdown-menu.tsx'
+export * from './icon.tsx'
+export * from './input.tsx'
+export * from './label.tsx'
+export * from './textarea.tsx'
+export * from './tooltip.tsx'
