+++ app/root.tsx
@@ -24,10 +24,10 @@ import { HoneypotProvider } from 'remix-utils/honeypot/react'
 import { z } from 'zod'
 import { Confetti } from './components/confetti.tsx'
 import { GeneralErrorBoundary } from './components/error-boundary.tsx'
+import { href as iconsHref } from './components/index.ts'
 import { EpicProgress } from './components/progress-bar.tsx'
 import { PageFooter, PageHeader } from './components/templates/index.ts'
 import { EpicToaster } from './components/toaster.tsx'
-import { href as iconsHref } from './components/ui/icon.tsx'
 import fontStyleSheetUrl from './styles/font.css'
 import tailwindStyleSheetUrl from './styles/tailwind.css'
 import { authenticator, getUserId } from './utils/auth.server.ts'
