+++ app/routes/_auth+/onboarding_.$provider.tsx
@@ -15,9 +15,13 @@ import {
 } from '@remix-run/react'
 import { safeRedirect } from 'remix-utils/safe-redirect'
 import { z } from 'zod'
-import { CheckboxField, ErrorList, Field } from '#app/components/forms.tsx'
-import { Spacer } from '#app/components/spacer.tsx'
-import { StatusButton } from '#app/components/ui/status-button.tsx'
+import {
+	CheckboxField,
+	ErrorList,
+	Field,
+	Spacer,
+	StatusButton,
+} from '#app/components/index.ts'
 import {
 	authenticator,
 	requireAnonymous,
