+++ app/routes/settings+/profile.change-email.tsx
@@ -6,9 +6,7 @@ import { json, redirect, type DataFunctionArgs } from '@remix-run/node'
 import { Form, useActionData, useLoaderData } from '@remix-run/react'
 import { AuthenticityTokenInput } from 'remix-utils/csrf/react'
 import { z } from 'zod'
-import { ErrorList, Field } from '#app/components/forms.tsx'
-import { Icon } from '#app/components/ui/icon.tsx'
-import { StatusButton } from '#app/components/ui/status-button.tsx'
+import { ErrorList, Field, Icon, StatusButton } from '#app/components/index.ts'
 import {
 	prepareVerification,
 	requireRecentVerification,
