+++ app/routes/settings+/profile.password.tsx
@@ -5,10 +5,13 @@ import { json, redirect, type DataFunctionArgs } from '@remix-run/node'
 import { Form, Link, useActionData } from '@remix-run/react'
 import { AuthenticityTokenInput } from 'remix-utils/csrf/react'
 import { z } from 'zod'
-import { ErrorList, Field } from '#app/components/forms.tsx'
-import { Button } from '#app/components/index.ts'
-import { Icon } from '#app/components/ui/icon.tsx'
-import { StatusButton } from '#app/components/ui/status-button.tsx'
+import {
+	Button,
+	ErrorList,
+	Field,
+	Icon,
+	StatusButton,
+} from '#app/components/index.ts'
 import {
 	getPasswordHash,
 	requireUserId,
