+++ app/routes/settings+/profile.photo.tsx
@@ -17,10 +17,7 @@ import {
 import { useState } from 'react'
 import { AuthenticityTokenInput } from 'remix-utils/csrf/react'
 import { z } from 'zod'
-import { ErrorList } from '#app/components/forms.tsx'
-import { Button } from '#app/components/index.ts'
-import { Icon } from '#app/components/ui/icon.tsx'
-import { StatusButton } from '#app/components/ui/status-button.tsx'
+import { Button, ErrorList, Icon, StatusButton } from '#app/components/index.ts'
 import { requireUserId } from '#app/utils/auth.server.ts'
 import { validateCSRF } from '#app/utils/csrf.server.ts'
 import { prisma } from '#app/utils/db.server.ts'
